requirements.txt
Flask
gunicorn
pandas
numpy
sklearn
PyYAML